db: rename polkadb to db
